Refactor `new-account.js` (#7312)
* Start refactor

* Use import syntax

* Add create-account.component

* Continue refactor

* Add new line

* Start using JSX and make tabs a bit more DRY

* :wave: bye-bye hyperscript

* These can be disabled when active

* Start JSX in new account component

* :wave: bye-bye hyperscript

* Move newAccountNumber into container

* Validate newAccountNumber prop
