+++ ui/app/css/itcss/components/new-account.scss
@@ -50,6 +50,7 @@
       color: $curious-blue;
       border-bottom: 3px solid $curious-blue;
       cursor: initial;
+      pointer-events: none;
     }
   }
 }
